+++ bot/cogs/admin.py
@@ -1,16 +1,16 @@
 from discord.ext import commands
 from logging import getLogger
+from datetime import datetime as dt
 
 import modules.config as cfg
 from modules.enumerations import SelStatus, MatchStatus, PlayerStatus
 from modules.display import send, channelSend
 from modules.exceptions import ElementNotFound, DatabaseError
-from modules.tools import isAdmin
 from modules.database import removePlayer as dbRemove
 from modules.loader import lockAll, unlockAll, isAllLocked
-from modules.roles import getRole, forceInfo
+from modules.roles import forceInfo, roleUpdate, isAdmin, permsMuted, channelFreeze
 
-from classes.players import removePlayer, getPlayer
+from classes.players import removePlayer, getPlayer, Player
 
 from matches import clearLobby, getMatch, getAllNamesInLobby, removeFromLobby
 
@@ -40,14 +40,14 @@ class AdminCog(commands.Cog, name='admin'):
     @commands.command()
     @commands.guild_only()
     async def clear(self, ctx):
-        if ctx.channel.id == cfg.discord_ids["lobby"]:  # clear lobby
+        if ctx.channel.id == cfg.channels["lobby"]:  # clear lobby
             if clearLobby():
                 await send("LB_CLEARED", ctx, namesInLobby=getAllNamesInLobby())
                 return
             await send("LB_EMPTY", ctx)
             return
         # clear a match channel
-        if ctx.channel.id in cfg.discord_ids["matches"]:
+        if ctx.channel.id in cfg.channels["matches"]:
             match = getMatch(ctx.channel.id)
             if match.status is MatchStatus.IS_FREE:
                 await send("MATCH_NO_MATCH", ctx, ctx.command.name)
@@ -65,8 +65,8 @@ class AdminCog(commands.Cog, name='admin'):
     @commands.guild_only()
     async def map(self, ctx, *args):
         channelId = ctx.channel.id
-        if channelId not in cfg.discord_ids["matches"]:
-            await send("WRONG_CHANNEL", ctx, ctx.command.name, " channels " + ", ".join(f'<#{id}>' for id in cfg.discord_ids["matches"]))
+        if channelId not in cfg.channels["matches"]:
+            await send("WRONG_CHANNEL", ctx, ctx.command.name, " channels " + ", ".join(f'<#{id}>' for id in cfg.channels["matches"]))
             return
         match = getMatch(ctx.channel.id)
         if match.status is MatchStatus.IS_FREE:
@@ -90,14 +90,14 @@ class AdminCog(commands.Cog, name='admin'):
             return
         if player.status is PlayerStatus.IS_LOBBIED:
             removeFromLobby(player)
-            await channelSend("RM_LOBBY", cfg.discord_ids["lobby"], player.mention, namesInLobby=getAllNamesInLobby())
+            await channelSend("RM_LOBBY", cfg.channels["lobby"], player.mention, namesInLobby=getAllNamesInLobby())
         if player.status in (PlayerStatus.IS_REGISTERED, PlayerStatus.IS_NOT_REGISTERED):
             try:
                 await dbRemove(player)
             except DatabaseError:
                 pass  # ignored if not yet in db
+            await forceInfo(player.id)
             removePlayer(player)
-            forceInfo(player)
             await send("RM_OK", ctx)
             return
         await send("RM_IN_MATCH", ctx)
@@ -110,31 +110,79 @@ class AdminCog(commands.Cog, name='admin'):
             return
         if player.status is PlayerStatus.IS_LOBBIED:
             removeFromLobby(player)
-            await channelSend("RM_LOBBY", cfg.discord_ids["lobby"], player.mention, namesInLobby=getAllNamesInLobby())
+            await channelSend("RM_LOBBY", cfg.channels["lobby"], player.mention, namesInLobby=getAllNamesInLobby())
         await send("RM_NOT_LOBBIED", ctx)
 
-    # WIP
-    # @commands.command()
-    # @commands.guild_only()
-    # async def timeout(self, ctx, *args):
-    #     if len(ctx.message.mentions) != 1:
-    #         await send("RM_MENTION_ONE", ctx)
-    #         return
-    #     try:
-    #         player = getPlayer(ctx.message.mentions[0].id)
-    #     except ElementNotFound:
-    #         # player isn't even registered in the system...
-    #         await send("RM_NOT_IN_DB", ctx)
-    #         return
-    #     if player.status is PlayerStatus.IS_LOBBIED:
-    #         removeFromLobby(player)
-    #         await channelSend("RM_LOBBY", cfg.discord_ids["lobby"], player.mention, namesInLobby=getAllNamesInLobby())
-    #     if player.status not in (PlayerStatus.IS_REGISTERED, PlayerStatus.IS_NOT_REGISTERED):
-    #         await send("RM_IN_MATCH", ctx)
-    #         return
+    @commands.command()
+    @commands.guild_only()
+    async def timeout(self, ctx, *args):
+        if len(args) == 0:
+            await send("RM_TIMEOUT_HELP", ctx)
+            return
+        if len(args) == 1 and args[0] == "help":
+            await send("RM_TIMEOUT_HELP", ctx)
+            return
+        if len(ctx.message.mentions) != 1:
+            await send("RM_MENTION_ONE", ctx)
+            return
+        try:
+            player = getPlayer(ctx.message.mentions[0].id)
+        except ElementNotFound:
+            # player isn't even registered in the system...
+            player = Player(ctx.message.mentions[0].id, ctx.message.mentions[0].name)
+            return
+        if player.status is PlayerStatus.IS_LOBBIED:
+            removeFromLobby(player)
+            await channelSend("RM_LOBBY", cfg.channels["lobby"], player.mention, namesInLobby=getAllNamesInLobby())
+        if player.status not in (PlayerStatus.IS_REGISTERED, PlayerStatus.IS_NOT_REGISTERED):
+            await send("RM_IN_MATCH", ctx)
+            return
+        if len(args) == 1:
+            if player.isTimeout:
+                await send("RM_TIMEOUT_INFO", ctx, dt.utcfromtimestamp(player.timeout).strftime("%Y-%m-%d %H:%M UTC"))
+                return
+            await roleUpdate(player)
+            await permsMuted(False, player.id)
+            await send("RM_TIMEOUT_NO", ctx)
+            return
+        # =timeout @player remove
+        if len(args) == 2 and args[1] == 'remove':
+            if not player.isTimeout:
+                await send("RM_TIMEOUT_ALREADY", ctx)
+                return
+            player.timeout = 0
+            await player.dbUpdate("timeout")
+            await send("RM_TIMEOUT_FREE", ctx, player.mention)
+            await roleUpdate(player)
+            await permsMuted(False, player.id)
+            return
+        # Check if command is correct (=timeout @player 12 d)
+        if len(args) != 3:
+            await send("RM_TIMEOUT_INVALID", ctx)
+            return
+        if args[2] in ['d', 'day', 'days']:
+            time = 86400
+        elif args[2] in ['h', 'hour', 'hours']:
+            time = 3600
+        elif args[2] in ['m', 'min', 'mins', 'minute', 'minutes'] :
+            time = 60
+        else:
+            await send("RM_TIMEOUT_INVALID", ctx)
+            return
+        try:
+            time *= int(args[1])
+            if time == 0:
+                raise ValueError
+        except ValueError:
+            await send("RM_TIMEOUT_INVALID", ctx)
+            return
+        endTime = int(dt.timestamp(dt.now()))+time
+        player.timeout = endTime
+        await roleUpdate(player)
+        await player.dbUpdate("timeout")
+        await permsMuted(True, player.id)
+        await send("RM_TIMEOUT", ctx, player.mention, dt.utcfromtimestamp(endTime).strftime("%Y-%m-%d %H:%M UTC"))
         
-    #     player.addTimeout
-    #     forceInfo(player)
 
 
 
@@ -169,24 +217,12 @@ class AdminCog(commands.Cog, name='admin'):
     async def channel(self, ctx, *args):
         if len(args) == 1:
             arg = args[0]
-            memb = ctx.author
-            notify = memb.guild.get_role(cfg.discord_ids["notify_role"])
-            registered = memb.guild.get_role(
-                cfg.discord_ids["registered_role"])
-            ov_notify = ctx.channel.overwrites_for(notify)
-            ov_registered = ctx.channel.overwrites_for(registered)
             if arg == "freeze":
-                ov_notify.send_messages = False
-                ov_registered.send_messages = False
-                await ctx.channel.set_permissions(notify, overwrite=ov_notify)
-                await ctx.channel.set_permissions(registered, overwrite=ov_registered)
+                await channelFreeze(True, ctx.channel.id)
                 await send("BOT_FROZEN", ctx)
                 return
             if arg == "unfreeze":
-                ov_notify.send_messages = True
-                ov_registered.send_messages = True
-                await ctx.channel.set_permissions(notify, overwrite=ov_notify)
-                await ctx.channel.set_permissions(registered, overwrite=ov_registered)
+                await channelFreeze(False, ctx.channel.id)
                 await send("BOT_UNFROZEN", ctx)
                 return
         await send("WRONG_USAGE", ctx, ctx.command.name)
@@ -196,8 +232,8 @@ def setup(client):
     client.add_cog(AdminCog(client))
 
 async def _removeChecks(ctx, channel):
-    if ctx.channel.id != cfg.discord_ids[channel]:
-        await send("WRONG_CHANNEL", ctx, ctx.command.name, f"<#{cfg.discord_ids[channel]}>")
+    if ctx.channel.id != cfg.channels[channel]:
+        await send("WRONG_CHANNEL", ctx, ctx.command.name, f"<#{cfg.channels[channel]}>")
         return
     if len(ctx.message.mentions) != 1:
         await send("RM_MENTION_ONE", ctx)